+++ src/test/java/org/apache/datasketches/hive/kll/GetRankUDFTest.java
@@ -19,11 +19,10 @@
 
 package org.apache.datasketches.hive.kll;
 
+import org.apache.datasketches.kll.KllFloatsSketch;
 import org.apache.hadoop.io.BytesWritable;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-
-import org.apache.datasketches.kll.KllFloatsSketch;
 
 @SuppressWarnings("javadoc")
 public class GetRankUDFTest {
@@ -43,7 +42,7 @@ public class GetRankUDFTest {
     sketch.update(4);
     final Double result = new GetRankUDF().evaluate(new BytesWritable(sketch.toByteArray()), 3f);
     Assert.assertNotNull(result);
-    Assert.assertEquals(result, 0.5);
+    Assert.assertEquals((double)result, 0.5);
   }
 
 }
