+++ src/test/java/org/apache/datasketches/hive/quantiles/GetCdfFromStringsSketchUDFTest.java
@@ -22,14 +22,12 @@ package org.apache.datasketches.hive.quantiles;
 import java.util.Comparator;
 import java.util.List;
 
-import org.apache.hadoop.io.BytesWritable;
-
-import org.testng.annotations.Test;
-import org.testng.Assert;
-
 import org.apache.datasketches.ArrayOfItemsSerDe;
 import org.apache.datasketches.ArrayOfStringsSerDe;
 import org.apache.datasketches.quantiles.ItemsSketch;
+import org.apache.hadoop.io.BytesWritable;
+import org.testng.Assert;
+import org.testng.annotations.Test;
 
 @SuppressWarnings("javadoc")
 public class GetCdfFromStringsSketchUDFTest {
@@ -52,7 +50,7 @@ public class GetCdfFromStringsSketchUDFTest {
     List<Double> result = new GetCdfFromStringsSketchUDF().evaluate(new BytesWritable(sketch.toByteArray(serDe)));
     Assert.assertNotNull(result);
     Assert.assertEquals(result.size(), 1);
-    Assert.assertEquals(result.get(0), 1.0);
+    Assert.assertEquals((double)result.get(0), 1.0);
   }
 
   @Test
@@ -71,11 +69,11 @@ public class GetCdfFromStringsSketchUDFTest {
     sketch.update("d");
     List<Double> result = new GetCdfFromStringsSketchUDF().evaluate(new BytesWritable(sketch.toByteArray(serDe)), "a", "c", "d");
     Assert.assertNotNull(result);
-    Assert.assertEquals(result.size(), 4);
-    Assert.assertEquals(result.get(0), 0.0);
-    Assert.assertEquals(result.get(1), 0.5);
-    Assert.assertEquals(result.get(2), 0.75);
-    Assert.assertEquals(result.get(3), 1.0);
+    Assert.assertEquals((double)result.size(), 4);
+    Assert.assertEquals((double)result.get(0), 0.0);
+    Assert.assertEquals((double)result.get(1), 0.5);
+    Assert.assertEquals((double)result.get(2), 0.75);
+    Assert.assertEquals((double)result.get(3), 1.0);
   }
 
 }
