+++ src/test/java/org/apache/datasketches/hive/quantiles/GetPmfFromStringsSketchUDFTest.java
@@ -22,14 +22,12 @@ package org.apache.datasketches.hive.quantiles;
 import java.util.Comparator;
 import java.util.List;
 
-import org.apache.hadoop.io.BytesWritable;
-
-import org.testng.annotations.Test;
-import org.testng.Assert;
-
 import org.apache.datasketches.ArrayOfItemsSerDe;
 import org.apache.datasketches.ArrayOfStringsSerDe;
 import org.apache.datasketches.quantiles.ItemsSketch;
+import org.apache.hadoop.io.BytesWritable;
+import org.testng.Assert;
+import org.testng.annotations.Test;
 
 @SuppressWarnings("javadoc")
 public class GetPmfFromStringsSketchUDFTest {
@@ -52,7 +50,7 @@ public class GetPmfFromStringsSketchUDFTest {
     List<Double> result = new GetPmfFromStringsSketchUDF().evaluate(new BytesWritable(sketch.toByteArray(serDe)));
     Assert.assertNotNull(result);
     Assert.assertEquals(result.size(), 1);
-    Assert.assertEquals(result.get(0), 1.0);
+    Assert.assertEquals((double)result.get(0), 1.0);
   }
 
   @Test
@@ -72,10 +70,10 @@ public class GetPmfFromStringsSketchUDFTest {
     List<Double> result = new GetPmfFromStringsSketchUDF().evaluate(new BytesWritable(sketch.toByteArray(serDe)), "a", "c", "e");
     Assert.assertNotNull(result);
     Assert.assertEquals(result.size(), 4);
-    Assert.assertEquals(result.get(0), 0.0);
-    Assert.assertEquals(result.get(1), 0.5);
-    Assert.assertEquals(result.get(2), 0.5);
-    Assert.assertEquals(result.get(3), 0.0);
+    Assert.assertEquals((double)result.get(0), 0.0);
+    Assert.assertEquals((double)result.get(1), 0.5);
+    Assert.assertEquals((double)result.get(2), 0.5);
+    Assert.assertEquals((double)result.get(3), 0.0);
   }
 
 }
