+++ src/test/java/org/apache/datasketches/hive/tuple/ArrayOfDoublesSketchToEstimatesUDFTest.java
@@ -21,12 +21,11 @@ package org.apache.datasketches.hive.tuple;
 
 import java.util.List;
 
+import org.apache.datasketches.tuple.ArrayOfDoublesUpdatableSketch;
+import org.apache.datasketches.tuple.ArrayOfDoublesUpdatableSketchBuilder;
 import org.apache.hadoop.io.BytesWritable;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-
-import org.apache.datasketches.tuple.ArrayOfDoublesUpdatableSketch;
-import org.apache.datasketches.tuple.ArrayOfDoublesUpdatableSketchBuilder;
 
 @SuppressWarnings("javadoc")
 public class ArrayOfDoublesSketchToEstimatesUDFTest {
@@ -43,8 +42,8 @@ public class ArrayOfDoublesSketchToEstimatesUDFTest {
     List<Double> result = new ArrayOfDoublesSketchToEstimatesUDF().evaluate(new BytesWritable(sketch.compact().toByteArray()));
     Assert.assertNotNull(result);
     Assert.assertEquals(result.size(), 2);
-    Assert.assertEquals(result.get(0), 0.0);
-    Assert.assertEquals(result.get(1), 0.0);
+    Assert.assertEquals((double)result.get(0), 0.0);
+    Assert.assertEquals((double)result.get(1), 0.0);
   }
 
   @Test
@@ -55,8 +54,8 @@ public class ArrayOfDoublesSketchToEstimatesUDFTest {
     List<Double> result = new ArrayOfDoublesSketchToEstimatesUDF().evaluate(new BytesWritable(sketch.compact().toByteArray()));
     Assert.assertNotNull(result);
     Assert.assertEquals(result.size(), 2);
-    Assert.assertEquals(result.get(0), 2.0);
-    Assert.assertEquals(result.get(1), 2.0);
+    Assert.assertEquals((double)result.get(0), 2.0);
+    Assert.assertEquals((double)result.get(1), 2.0);
   }
 
 }
